Use prebuilt module so Windows users will not have to rebuild the package
Use prebuilt module so Windows users will not have to rebuild the package on their own. The binary-test.js file is used to determine if the built binary is correct. As I'm not sure how to test the code and as I know the binaries are correct, I just write "pass" - this indicates to prebuilt module that everything is correct.